Move modal logic to header.
